Create sendingDataToTheMqttServer.ino
Using the 12-pin nodeMcu board and the arduino IDE, a second code was developed that monitors limit switches and sends data to a local MQTT server via wireless internet. In addition, the same code collects information online to know the current time according to the online network without the need for an rtc component like computers use.